+++ schema/schema.go
@@ -17,8 +17,6 @@ package schema
 import (
 	"context"
 	"fmt"
-	sdk "github.com/conduitio/conduit-connector-sdk"
-
 	"github.com/conduitio/conduit-commons/schema"
 	"github.com/conduitio/conduit-connector-protocol/pconduit"
 	"github.com/conduitio/conduit-connector-protocol/pconduit/v1/client"
@@ -50,7 +48,12 @@ func Create(ctx context.Context, typ schema.Type, subject string, bytes []byte) 
 
 func qualifiedSubject(ctx context.Context, subject string) string {
 	// todo better delimiter
-	return fmt.Sprintf("%s_%s", sdk.GetSchemaContextName(ctx), subject)
+	schemaCtx := GetSchemaContextName(ctx)
+	if schemaCtx == "" {
+		return subject
+	}
+
+	return fmt.Sprintf("%s_%s", schemaCtx, subject)
 }
 
 // Get retrieves the schema with the given name and version. If the schema does not exist, an error is returned.
@@ -65,6 +68,25 @@ func Get(ctx context.Context, subject string, version int) (schema.Schema, error
 
 	return resp.Schema, nil
 }
+
+// -- Schema context utilities -------------------------------------------------
+
+type schemaContextNameKey struct{}
+
+func WithSchemaContextName(ctx context.Context, schemaCtxName string) context.Context {
+	return context.WithValue(ctx, schemaContextNameKey{}, schemaCtxName)
+}
+
+func GetSchemaContextName(ctx context.Context) string {
+	name := ctx.Value(schemaContextNameKey{})
+	if name != nil {
+		return name.(string) //nolint:forcetypeassert // only this package can set the value, it has to be a string
+	}
+
+	return ""
+}
+
+// -- Schema service initializer -----------------------------------------------
 
 type standaloneInitializer struct{}
